refactor(lessons): KnowledgeCheckPlayer sous 300 lignes (H5)
Éclatement complet (pattern G1) : 970 → 76 lignes (orchestrateur).
- useKnowledgeCheckPlayer.js (159 l) : machine à états intro→playing→results,
  tentative, minuteur, réponses, navigation, soumission. Logique verbatim.
- QuizPlayerIntro.vue (187 l) : écran d'intro (stats, meilleur score, démarrer).
- QuizPlayerQuestion.vue (216 l) : question en cours (progression, minuteur,
  options) ; helpers d'affichage purs locaux.
- QuizQuestionNav.vue (170 l) : navigation (précédent/pastilles/suivant/terminer).
- QuizPlayerResults.vue (219 l) : bandeau résultat + score + actions.
- QuizPlayerCorrections.vue (131 l) : corrections par question (getAnswerText).
- Tests : composable (6) + 5 montages (21). CSS déplacé verbatim, @media (640px)
  réparti selon l'élément ciblé. Aucune dette pré-existante modifiée.
